edit field

+++ workspace/ims-webapp/src/main/java/th/go/excise/ims/ta/persistence/repository/TaDraftWorksheetHdrRepository.java
@@ -1,9 +1,8 @@
 package th.go.excise.ims.ta.persistence.repository;
 
-import java.math.BigDecimal;
 import th.co.baiwa.buckwaframework.common.persistence.repository.CommonJpaCrudRepository;
 import th.go.excise.ims.ta.persistence.entity.TaDraftWorksheetHdr;
 
-public interface TaDraftWorksheetHdrRepository extends CommonJpaCrudRepository<TaDraftWorksheetHdr, BigDecimal> {
+public interface TaDraftWorksheetHdrRepository extends CommonJpaCrudRepository<TaDraftWorksheetHdr, Long> {
 
 }
